Reinstate formatter accceptance tests
