+++ test/support/rake_task_definition.rb
@@ -1,6 +1,7 @@
 module RakeTaskDefinition
   def define_and_execute_rake_task(task_name, &block)
     task_name ||= RakeTaskDefinition.unique_task_name(name)
+    Rake::Task[task_name].clear if Rake::Task.task_defined?(task_name)
     Rake::Task.define_task(task_name, &block).execute
   end
